Added background colour key to the power page.
As a test of my ColourID scripts, and to make sure I could appropriately
describe the process of adding ColourID values to classes, I added a ColourId
to Page::Power that can be used to set the page background color. No specific
background color is defined by default, so the power page still uses the
default background color if you don't specify a new one in colours.json.