Documented ColourIds, adjusted ColourId scripts.
A file documenting the full process for adding and using ColourId values is now
provided. While writing this document, the whole process was tested, and
several bugs in the ColourID scripts were fixed. For more convenient editing of
ColourID Elements and namespaces, their menus no longer close automatically when
changes are made.