Using safer asynchronous message thread calls.
All components that need to be updated asynchronously on the message thread now
use juce::Component::SafePointer objects to ensure the asynchronous updates are
cancelled if the component is destroyed. Additionally, DrawableImage now always
locks the message thread, so image changes no longer need to be called on the
message thread unless it would risk a deadlock.